pyproject.toml
[project]
name = "leafdiseaseclassification"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "pandas>=2.2.3",
    "torch>=2.5.1",
    "torchvision>=0.20.1",
    "tqdm>=4.67.1",
    "scikit-learn>=1.6.0",
    "notebook>=7.3.2",
    "wandb>=0.19.1",
    "python-dotenv>=1.0.1",
]
